// src/sales/dto/create-sale.dto.ts
export class CreateSaleDto {
  idSystemGc: number;

  idAzatGc: number;

  idUserGc: number;

  userName: string;

  userEmail: string;

  createdAt: string;

  userPhone: string;

  orderName: string;

  payedAt: string;

  price: number;

  dealStatus: string;

  payFee: number;

  payedPrice: number;

  taxes: number;

  income: number;

  otherFee: number;

  profit: number;

  managerName: string;

  netProfit: number;

  payedBy: string;

  city: string;

  promoCompany: string;

  promocodeUsed: string;

  utmMedium: string;

  utmSource: string;

  utmContent: string;

  utmCampaign: string;

  utmGroup: string;

  utmTerm: string;

  orderComments: string;

  workWithOrder: string;

  orderTag: string;

  rejectReason: string;
}